Create LeetCode_2528.cpp
Nice question. There is a standard way to solve this kind of question. I am thinking to use greedy to solve it. But again, too much to think about it. Just use binary search and verify the guess is very efficient. So you need to be aware of the method...